python-server/requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
PyPDF2
langchain
langchain-google-genai
langchain_community
python-multipart
faiss-cpu
pymupdf
groq
livekit-agents[google]~=1.0
livekit-plugins-google
livekit-agents[images]
livekit-agents[bey]~=1.0
livekit-plugins-bey